% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlogram_Plot_of_Num_Var_Function.R
\name{corr_plot}
\alias{corr_plot}
\title{Correlogram Plot of Num. Var. Function}
\usage{
corr_plot(x)
}
\arguments{
\item{x}{The year you select in the dataset}
}
\value{
The output from \code{\link{corrplot}}, which is a correlogram plot
}
\description{
This function allows you to obtain the correlogram plot of the numeric variables in an NBA Season Stats Dataset, given a selected year
}
\examples{
\dontrun{
corr_plot(2000)
}

}
\keyword{Correlogram}
