% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Name_of_First_Oldest_Player_Function.R
\name{first_player_age}
\alias{first_player_age}
\title{Name of First Oldest Player Function}
\usage{
first_player_age(x)
}
\arguments{
\item{x}{The year you select in the dataset}
}
\value{
The output from \code{\link{print}}, which is the name of that first listed player
}
\description{
This function allows you to obtain name of the first player listed among the oldest players in an NBA Season Stats Dataset, given a selected year
}
\examples{
\dontrun{
first_player_age(2000)
}

}
\keyword{Oldest}
