% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First_Player_Position_Function.R
\name{first_player_pos}
\alias{first_player_pos}
\title{First Player Position Function}
\usage{
first_player_pos(x)
}
\arguments{
\item{x}{The year you select in the dataset}
}
\value{
The output from \code{\link{print}}, which is the position of that first listed player
}
\description{
This function allows you to obtain the position of the first player listed in an NBA Season Stats Dataset, given a selected year
}
\examples{
\dontrun{
first_player_pos(2000)
}

}
\keyword{Position}
